requirements.txt
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
boto3==1.26.158
botocore==1.29.158
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
comm==0.1.3
comtypes==1.1.14
coreapi==2.3.3
coreschema==0.0.4
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
Django==4.1.5
django-cors-headers==3.14.0
django-filter==22.1
django-phonenumber-field==7.0.2
django-phonenumbers==1.0.1
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djongo==1.3.6
dnspython==2.3.0
drf-yasg==1.21.4
executing==1.2.0
fastjsonschema==2.16.3
filelock==3.9.0
Flask==2.2.3
fqdn==1.5.1
idna==3.4
inflection==0.5.1
ipykernel==6.22.0
ipython==8.12.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
itsdangerous==2.1.2
itypes==1.2.0
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mistune==2.0.5
mouritech-mail-notification==0.0.2
nbclassic==0.5.5
nbclient==0.7.3
nbconvert==7.3.0
nbformat==5.8.0
nest-asyncio==1.5.6
notebook==6.5.4
notebook_shim==0.2.2
openapi-codec==1.3.2
packaging==23.0
pandocfilters==1.5.0
parso==0.8.3
pdfkit==1.0.0
phonenumbers==8.13.7
pickleshare==0.7.5
pika==1.3.2
platformdirs==2.6.2
prometheus-client==0.16.0
prompt-toolkit==3.0.38
psutil==5.9.4
psycopg2==2.9.6
pure-eval==0.2.2
pycparser==2.21
Pygments==2.14.0
PyJWT==2.6.0
pymongo==3.12.3
pypiwin32==223
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.7
pyttsx3==2.90
pytz==2022.7.1
pywin32==306
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.2
QtPy==2.3.1
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
ruamel.yaml==0.17.21
s3transfer==0.6.1
Send2Trash==1.8.0
simplejson==3.18.1
six==1.16.0
websocket-client==1.5.1
Werkzeug==2.2.3
widgetsnbextension==4.0.7
wkhtmltopdf==0.2